+++ core/src/main/scala/info/mukel/telegrambot4s/methods/EditMessageText.scala
@@ -16,7 +16,7 @@ import info.mukel.telegrambot4s.models.{ChatId, Message, ReplyMarkup}
   */
 case class EditMessageText(
                           chatId                : Option[ChatId] = None,
-                          messageId             : Option[Long] = None,
+                          messageId             : Option[Int] = None,
                           inlineMessageId       : Option[String] = None,
                           text                  : String,
                           parseMode             : Option[ParseMode] = None,
