+++ core/src/main/scala/info/mukel/telegrambot4s/methods/SendVenue.scala
@@ -26,6 +26,6 @@ case class SendVenue(
                     foursquareId        : Option[String] = None,
                     duration            : Option[String] = None,
                     disableNotification : Option[Boolean] = None,
-                    replyToMessageId    : Option[Long] = None,
+                    replyToMessageId    : Option[Int] = None,
                     replyMarkup         : Option[ReplyMarkup] = None
                   ) extends ApiRequestJson[Message]
