+++ core/src/main/scala/info/mukel/telegrambot4s/methods/SendVoice.scala
@@ -26,6 +26,6 @@ case class SendVoice(
                       caption             : Option[String] = None,
                       duration            : Option[Int] = None,
                       disableNotification : Option[Boolean] = None,
-                      replyToMessageId    : Option[Long] = None,
+                      replyToMessageId    : Option[Int] = None,
                       replyMarkup         : Option[ReplyMarkup] = None
                     ) extends ApiRequestMultipart[Message]
