+++ core/src/main/scala/info/mukel/telegrambot4s/models/Contact.scala
@@ -11,5 +11,5 @@ case class Contact(
                     phoneNumber : String,
                     firstName   : String,
                     lastName    : Option[String] = None,
-                    userId      : Option[Long] = None
+                    userId      : Option[Int] = None
                   )
